backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
flask-mail==0.9.1
supabase==2.3.0
python-dotenv==1.0.0
reportlab==4.0.8
pytest==7.4.3
requests==2.31.0
gunicorn==21.2.0
mysql-connector-python==8.2.0
PyJWT==2.8.0